Memory deallocation
Introduced a mechanism for memory deallocation. Particularly subopt can return
a large number of results (occupying a big portion of memory) so all c
allocated memory must be freed.

When returning a c_char_p, ctypes automatically converts it to a string. This
prevents the possibility of freeing it. An alternative is to return a c_void_p
and casting it to a string when needed, keeping around the c_void_p for
deallocation.