New functions for testing compatibility of sequence to structure
When calling seq_eval the sequence and structure being tested might not be \
compatible (i.e. structure indicates a pair but there is an A and a C in the \
sequence) meaning the return energy of the fold is not meaningful. A new \
function seq_str_compatible checks that all the pairings of a structure \
(returned by str_base_pairs) are canonical pairing (plus the wobble).

Update version number.